pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lora-captioner"
version = "0.1.0"
description = "Command-line tool to caption images for LoRA training datasets using Vision Language Models"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "jmpijll" }
]
keywords = [
    "lora",
    "stable-diffusion",
    "flux",
    "captioning",
    "image-captioning",
    "vllm",
    "machine-learning",
    "dataset"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "transformers>=4.36.0",
    "torch>=2.0.0",
    "Pillow>=10.0.0",
    "click>=8.0.0",
    "tqdm>=4.65.0",
    "huggingface-hub>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
lora-captioner = "lora_captioner.cli:main"

[project.urls]
Homepage = "https://github.com/jmpijll/lora-captioner"
Repository = "https://github.com/jmpijll/lora-captioner"
Issues = "https://github.com/jmpijll/lora-captioner/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/lora_captioner"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
